requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
loguru==0.7.2
arxiv==2.1.3
beautifulsoup4==4.12.3
httpx==0.27.2
openai-whisper==20240930
faster-whisper==1.0.3
soundfile==0.12.1
transformers==4.44.2
torch>=2.3.1
pyttsx3==2.90
